requirements-windows.txt
tzlocal==1.4
asn1crypto==0.22.0
cffi==1.10.0
cryptography==1.8.1
idna==2.5
packaging==16.8
paramiko==2.1.2
pyasn1==0.2.3
pycparser==2.17
pyparsing==2.2.0
pytz==2017.2
PyYAML==3.12
six==1.10.0
langcodes==1.4.0
path.py==11.0
psutil==5.4.5
humanfriendly==4.12
requests==2.18.4
